remove gripper, implement perfect grasp, add test code for three-box-stacking
